# factory-saas/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 厂区访客管控 SaaS API
  description: |
    Factory Visitor Control SaaS — 多租户访客手机管控系统 API。
    
    ## 认证方式
    - Cookie 认证：登录后通过 `token` HttpOnly cookie 进行身份验证
    - CSRF 保护：非安全方法（POST/PUT/DELETE）需携带 `X-CSRF-Token` 头，值与 `csrf_token` cookie 一致
    
    ## 错误格式
    所有 API 错误返回统一格式：
    ```json
    {
      "error": "ERROR_CODE",
      "message": "用户可见描述",
      "timestamp": "2024-01-01T00:00:00.000Z",
      "requestId": "uuid"
    }
    ```
  version: 2.0.0
  contact:
    name: Factory SaaS Support
servers:
  - url: http://localhost:3000
    description: 开发环境
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: 请求参数无效
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        isSuperAdmin:
          type: boolean
...
  parameters:
    csrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: 用户登录
      description: 验证邮箱密码，成功后设置 token cookie
      parameters:
        - $ref: '#/components/parameters/csrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
                  redirect:
                    type: string
        '401':
          description: 凭据无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 登录过于频繁
...
  /api/auth/logout:
    post:
      tags: [auth]
      summary: 登出
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 登出成功
  /api/auth/me:
    get:
      tags: [auth]
      summary: 获取当前用户信息
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 当前用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: 未登录

  # ── Checkin ─────────────────────────────────────────────────
  /api/checkin-qr:
    get:
      tags: [checkin]
      summary: 生成签到二维码
      security:
        - cookieAuth: []
      parameters:
        - name: area
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 二维码数据
          content:
            application/json:
              schema:
                type: object
                properties:
                  qr:
                    type: string
                    description: 二维码 DataURL
                  url:
                    type: string
                  siteId:
                    type: string
                  subscriptionId:
                    type: string
                  checkinToken:
                    type: string
...
  /api/checkin:
    post:
      tags: [checkin]
      summary: 访客签到
      parameters:
        - $ref: '#/components/parameters/csrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, area]
              properties:
                name:
                  type: string
                company:
                  type: string
                area:
                  type: string
                siteId:
                  type: string
                checkinToken:
                  type: string
      responses:
        '200':
          description: 签到成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  entryQR:
                    type: string
                  exitQR:
                    type: string
        '429':
          description: 签到过于频繁

  # ── Sessions ────────────────────────────────────────────────
...